fix(vibevoice-omni-bench): default the CUDA host to float32; record cross-arm FLEURS parity
The Python host defaulted to bfloat16. On asr.fleurs.en.quick.v1 that put the
reference arm at wer_norm 0.781 against the MLX port's 0.581 -- a 0.20 gap that
reads as a broken port.

It was the host. The checkpoint stores every tensor as BF16 but declares
torch_dtype float32 at the top level and bfloat16 only for the decoder;
model.to(bfloat16) ignores that split and downcasts the acoustic and semantic
VAE encoders as well. MLX keeps the stored BF16 weights but type-promotes them
against the fp32 audio activations, so the port's effective compute precision
tracks the fp32 PyTorch run rather than the bf16 one.

  arm             wer_norm   hypotheses identical to the port
  cuda bf16       0.7812     18/64
  cuda fp32       0.5836     39/64
  swift mlx       0.5807     --

Defaulting to float32 is both what the checkpoint asks for and what makes the
arms comparable.

CROSS-ARM RESULT on asr.fleurs.en.quick.v1, 64/64 samples on each side:

  quality.wer_norm.v1   0.5836 -> 0.5807   delta -0.0028   GATE PASS (abs_lte 0.005)
  quality.cer.v1        0.4594 -> 0.4577   delta -0.0018
  quality.wer_ortho.v1  0.6992 -> 0.7034   delta +0.0042
  latency p50           1.54s  -> 2.01s
  rtfx                  6.21   -> 4.77
  peak rss gb           13.31  -> 4.63

The quality gate passes: on real speech the port matches the reference to
within 0.3% WER.

parity.hypothesis_identity.v1 is 0.64 rather than 1.0 -- 39 of 64 hypotheses
are byte-identical. Greedy decoding turns a near-tie argmax flipped by the last
bits of a different matmul kernel into one changed token, and the
autoregressive context carries it. Token-for-token equality was established
earlier on a fixed multi-chunk clip; it does not survive 64 clips times
hundreds of steps on different silicon. The aggregate quality metric is the
one that answers the porting question, and it agrees.

The overall parity verdict stays fail on identity_scope_mismatch (os is not an
allowed difference axis), so the gate evaluations have to be read directly.

+++ tools/omni_bench_host_vibevoice/vibevoice_host.py
@@ -466,7 +466,18 @@ def make():
     if not model_path:
         raise RuntimeError("VIBEVOICE_MODEL_PATH is required")
 
-    dtype_name = os.environ.get("VIBEVOICE_DTYPE", "bfloat16")
+    # float32 by default, matching the checkpoint's own top-level
+    # `torch_dtype: float32`. Forcing bfloat16 across the whole graph is not a
+    # free speedup here: on asr.fleurs.en.quick.v1 it moved wer_norm from
+    # 0.584 to 0.781 (measured), because `model.to(bfloat16)` also downcasts
+    # the acoustic/semantic VAE encoders that the config keeps in fp32.
+    #
+    # It also makes this arm incomparable to the MLX arm. MLX keeps the bf16
+    # weights as stored but type-promotes against the fp32 audio activations,
+    # so its effective compute precision tracks the fp32 PyTorch run, not the
+    # bf16 one. Defaulting to bfloat16 would therefore charge the port with a
+    # ~0.20 WER gap that belongs to this host's dtype handling.
+    dtype_name = os.environ.get("VIBEVOICE_DTYPE", "float32")
     dtype = {"float32": torch.float32, "bfloat16": torch.bfloat16,
              "float16": torch.float16}[dtype_name]
